PPL Zpl labels
