+++ src/Exception/NotImplementedException.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * User: Martinus - Samuel Szabo
+ * Date: 24.11.2017
+ */
+
+namespace Salamek\PplMyApi\Exception;
+
+
+class NotImplementedException extends \BadMethodCallException
+{
+
+}
