+++ test/PublicTest.php
@@ -7,7 +7,8 @@
 use Salamek\PplMyApi\Enum\LabelDecomposition;
 use Salamek\PplMyApi\Model\Package;
 use Salamek\PplMyApi\Tools;
-use Salamek\PplMyApi\Label;
+use Salamek\PplMyApi\PdfLabel;
+use Salamek\PplMyApi\ZplLabel;
 
 final class PublicTest extends BaseTest
 {
@@ -33,6 +34,14 @@ final class PublicTest extends BaseTest
     public function testPackageNumber()
     {
         $this->assertEquals('40950000115', $this->package->getPackageNumber());
+    }
+
+    /**
+     * @test
+     */
+    public function testPackageNumberChecksum()
+    {
+        $this->assertEquals('7', $this->package->getPackageNumberChecksum());
     }
 
     /**
@@ -83,7 +92,7 @@ final class PublicTest extends BaseTest
      */
     public function testGeneratePdfFullSinglePackage()
     {
-        $raw = Label::generateLabels([$this->package], LabelDecomposition::FULL);
+        $raw = PdfLabel::generateLabels([$this->package], LabelDecomposition::FULL);
 
         $this->assertNotEmpty($raw);
 
@@ -99,7 +108,7 @@ final class PublicTest extends BaseTest
      */
     public function testGeneratePdfFullMultiplePackages()
     {
-        $raw = Label::generateLabels($this->packages, LabelDecomposition::FULL);
+        $raw = PdfLabel::generateLabels($this->packages, LabelDecomposition::FULL);
 
         $this->assertNotEmpty($raw);
 
@@ -122,7 +131,7 @@ final class PublicTest extends BaseTest
      */
     public function testGeneratePdfQuarterSinglePackage()
     {
-        $raw = Label::generateLabels([$this->package], LabelDecomposition::QUARTER);
+        $raw = PdfLabel::generateLabels([$this->package], LabelDecomposition::QUARTER);
 
         $this->assertNotEmpty($raw);
 
@@ -138,7 +147,7 @@ final class PublicTest extends BaseTest
      */
     public function testGeneratePdfQuarterMultiplePackages()
     {
-        $raw = Label::generateLabels($this->packages, LabelDecomposition::QUARTER);
+        $raw = PdfLabel::generateLabels($this->packages, LabelDecomposition::QUARTER);
 
         $this->assertNotEmpty($raw);
 
@@ -154,5 +163,18 @@ final class PublicTest extends BaseTest
         file_put_contents($filePath, $raw);
 
         $this->assertFileExists($filePath);
+    }
+
+    /**
+     * @test
+     */
+    public function testGenerateZplLabel()
+    {
+        $zplString = ZplLabel::generateLabels([$this->package]);
+        $this->assertNotEmpty($zplString);
+        $this->assertEquals(
+            '^XA^MUM^LH2,2^FS^LL^PW^PON^FO47,49^LRY^GB48,80,0.3,B,0^FS^FO13,49^LRY^GB28,80,0.3,B,0^FS^FO4.7,114^LRY^GB6,15,0.3,B,0^FS^FO47.2,107.7^LRY^GB21,21,10.5,B,0^FS^FO88,7^GFA,2520,2520,9,00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004000000000000000007C00000000000000007F80000000000000001FE0000000000000000FE0000000000000000E60000000000000000E601E00000000000006601FE0000000000007E01FFC000000000043E01FFC40000000007C001FFC78000000007F801FFC7F800000000FE01FFC7FC000000007E01FFC7FCE00000006601FFC7FCFE0000046601FFC7FCFE000007E601FFC7FCFE380007FE01FFC7FCFE380000BE01FFC7FCFE3800001E01FFC7FCFE380003E001FFC7FCFE380007F801FFC7FCFE38000E3C01FFC7FCFE38000E0E01FFC7FCFE38000E0601FFC7FCFE3800060601FFC7FCFE3800070701FFC7FCFE380003C601FFC7FCFE380001FE01FFC7FCFE380000FC01FFC7FCFE3800070001FFC7FCFE380007F001FFC7FCFE380003FE01FFC7FCFE3800007E01FFC7FCFE3800006601FFC7FCFE3800006601FFC7FCFE3800006601FFC7FCFE3800006601FFC7FCFE3800070601FFC7FCFE380007E601FFC7FCFE380007FE01FFC7FCFE3800067E01FFC7FCFE3800066601FFC7FCFE3800066601FFC7FCFE3800066601FFC7FCFE3800066601FFC7FCFE3800002601FFC7FCFE3800070601FFC7FCFE3800073801FFC7FCFE38000E7C01FFC7FCFE38000E7E01FFC7FCFE38000E66003FC7FCFE380006670007C7FCFE380007E7000047FCFE380003EE000007FCFE380001CE000000FCFE3800070C0000001CFE3800073000000000FE38000E7C000000003E38000E7E000000000638000E6600000000003800066701E0000000080007E701FC000000000003EE01FFC00000000001CE01FFF800000000070C01FFFF8000000007F001FFFFF000000001FE01FFFFFF000000003E01FFFFFFE0000003E201FFFFFFFE000007F801FFFFFFFFC000077C01FFFFFFFFF8000E0E01FFFFFFFFF8000E0601FFFFFFFFF800060601FFFFFFFFF800070701FFFFFFFFF800038701FFFFFFFFF80001FE01FFFFFFFFF80000FE01FFFFFFFFF800071801FFFFFFFFF80007F001FFFFFFFFF80001FE007FFFFFFFF800003E0007FFFFFFF800007E0000FFFFFFF80003FC00000FFFFFF80007C0000007FFFFF80007E0000007FFFFF80003FC000007FBFFF800047E000007F83FF8000706000007F807F8000780000007F803F80001E0000007F803F80001F8000007F803F80001BC000007F803F800019E000007F803F80007FE000007F803F80007FE000007F803F8000780000007F803F8000400000007F803F80007C0000007F803F80007F8000007F803F80006FE000007F803F800061E000007F803F8000600000007F803F8000600000007F803F8000600000007FC07F8000000000007FE07F8000000000007FF0FF8000000000007FFFFF8000000000003FFFFF8000000000003FFFFF8000000000003FFFFF8000000000003FFFFF8000000000003FFFFF8000400000001FFFFF8000780000001FFFFF80007F8000001FFFFF80001FE000000FFFFF00000FE000000FFFFF00000E60000007FFFF00000E60000007FFFE00000660000003FFFE000007E0000001FFFC000063E0000000FFF8000070C00000007FF800007C000000003FE000001E000000000FC000001F80080000000000001BC01F00000000000018E01FF000000000007FE01FFE00000000007FE01FFFC00000000040001FFFFC0000000040001FFFFF800000007C001FFFFFF80000007F801FFFFFFF0000000FE01FFFFFFFF0000006E01FFFFFFFFE000006601FFFFFFFFF800066601FFFFFFFFF80007E601FFFFFFFFF80007FE01FFFFFFFFF800003E01FFFFFFFFF800001E01FFFFFFFFF80007F001FFFFFFFFF80007F801FFFFFFFFF8000E3C01FFFFFFFFF8000E0E00FFFFFFFFF8000606003FFFFFFFF80006070003FFFFFFF800070700007FFFFFF800038E000007FFFFF800001E000007FFFFF800070C000007FFFFF80007F0000007F9FFF80007FE000007F81FF800067E000007F803F8000666000007F803F8000666000007F803F8000666000007F803F8000666000007F803F8000026000007F803F8000706000007F803F80007F0000007F803F80007FE000007F803F800063E000007F803F8000606000007F803F8000600000007F803F8000600000007F803F8000200000007F803F8000000000007F803F8000000000007FC03F8000000000007FC07F8000000000007FE07F8000000000007FF9FF8000000000007FFFFF8000000000003FFFFF8000000000003FFFFF8000700000003FFFFF80007E0000003FFFFF80007FE000003FFFFF800067E000001FFFFF800060E000001FFFFF8000600000000FFFFF8000600000000FFFFF0000600000000FFFFF00000000000007FFFF00003F00000003FFFE00007F80000003FFFE000073C0000001FFFC0000E0E0000000FFF80000E0600000007FF0000060600000003FE000007070000000078000003C701C0000000000001FE01F8000000000000FE01FF800000000003E001FFF00000000007F001FFFF0000000007F801FFFFE00000000E1C01FFFFFE0000000E0E01FFFFFFC00000060601FFFFFFFC000006E701FFFFFFFF800007E701FFFFFFFFF00007EE01FFFFFFFFF80001EE01FFFFFFFFF800060C01FFFFFFFFF80007C001FFFFFFFFF80007FC01FFFFFFFFF800007E01FFFFFFFFF800070E01FFFFFFFFF800073801FFFFFFFFF8000E7C01FFFFFFFFF8000E7E01FFFFFFFFF8000E6601FFFFFFFFF800066701FFFFFFFFF80007E701FF9FFFFFF80003EE01FF83FFFFF80001CE01FF807FFFF800000C01FF8007FFF800000601FF8000FFF800070601FF80000FF80007E601FF800001F80003FE01FF8000001800007E01FF8000000000000601FF8000000000000601FF8000000000070601FF800000000007F601FF800000000003FE01FF8000000000003E01FF800000000003E601FF800000000007F801FF80000000000FFC01FF80000000000E0E01FF80000000000E0601FF8000000000060601FF8000000000070701FF8000000000038F01FF8000000000019E01FF8000000000000C01FF8000000000000001FF8000000000000001FF8000000000000001FF8000000000000000FF80000000000000001F80000000000000000380000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001F00000000000000007FC000000000000000C060000000000000014010000000000000037818000000000000063F0C0000000000000607EC0000000000000404E4000000000000040664000000000000040664000000000000046664000000000000067E6C000000000000020BE80000000000000301D800000000000001803000000000000000C060000000000000007FC0000000000000000E0000000000000000000000^FS^FO24,35^A0N,6,6^FD40950000115^FS^FO95,49^A0R,2,2^FDConsignee / Prijemce^FS^FO41,49^A0R,2,2^FDConsignor / Odesilatel^FS^FO86,51^AFR,4^FDAdam Schubert^FS^FO81,51^AFR,4^FDNa tabulovem vrchu 7^FS^FO76,51^AFR,4^FDOlomouc, CZ^FS^FO60,51^A0R,8,8^FD77900^FS^FO55,51^A0R,4,4^FDkontakt^FS^FO50,51^A0R,3,3^FDTEL.: ^FS^FO50,59^A0R,3,3^FD+420123456789^FS^FO53.7,110^A0R,7,7^FDDen^FS^FO35,51^AFR,4^FDMy Compamy s.r.o.^FS^FO30,51^AFR,4^FDMy Address^FS^FO25,51^AFR,4^FDOlomouc, CZ^FS^FO16,51^A0R,4,4^FD77900^FS^FO5,118^A0R,5,5^FD1/1^FS^FO80,7^ACR,2^FDmodra linka : 844 775 775^FS^FO76,7^ACR,2^FDhttps://www.ppl.cz^FS^FO72,7^ACR,2^FDinfo@ppl.cz^FS^FO66,7^ACR,2^FD^FS^FO8,49^A0R,2,2^FD^FS^FO3,49^A0R,2,2^FD^FS^FO20,2^BY0.40,3,32^B2N,32,N,N^FD409500001157^FS^XZ',
+            $zplString
+        );
     }
 }
